third_party/WebKit/LayoutTests/fast/dom/shadow/input-color-in-content-expected.txt

CONSOLE WARNING: line 13: Element.createShadowRoot is deprecated and will be removed in M73, around March 2019. Please use Element.attachShadow instead. See https://www.chromestatus.com/features/4507242028072960 for more details.
PASS mockColorChooser.isChooserShown() is false
PASS mockColorChooser.isChooserShown() is true
PASS mockColorChooser.isChooserShown() is true
PASS mockColorChooser.isChooserShown() is true
PASS mockColorChooser.isChooserShown() is true
PASS mockColorChooser.isChooserShown() is true
PASS mockColorChooser.isChooserShown() is true
PASS mockColorChooser.isChooserShown() is true
PASS mockColorChooser.isChooserShown() is true
PASS mockColorChooser.isChooserShown() is true
PASS mockColorChooser.isChooserShown() is true
PASS mockColorChooser.isChooserShown() is true
PASS successfullyParsed is true

TEST COMPLETE
